requirements.txt
asgiref==3.8.1
Django==5.1.5
django-environ==0.12.0
django-markdownx==4.0.7
Markdown==3.7
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
